requirements.txt
pandas
python-dotenv
requests
gigachain
langchain_gigachat==0.3.12
langchain==0.3.27
langfuse

# Telegram Bot
pyTelegramBotAPI>=4.14.0
